Join two data frames using a Bloom filter implemented with Rcpp

This function uses a high-performance Rcpp implementation of Bloom filters to efficiently join large data frames. The Bloom filter is used to pre-filter the left data frame, reducing the number of comparisons needed for the actual join.

Usage

bloom_join(
  x,
  y,
  by = NULL,
  type = "inner",
  bloom_size = NULL,
  false_positive_rate = 0.01
)

Arguments

x

A data frame

y

A data frame

by

A character vector of variables to join by

type

Type of join: "inner", "left", "right", "full", "semi", or "anti"

bloom_size

Size of the Bloom filter. If NULL, will be automatically calculated

false_positive_rate

False positive rate for the Bloom filter (default: 0.01)

Value

A joined data frame with a class of "bloomjoin"

Examples

x <- data.frame(id = 1:100000, value_x = rnorm(100000))
y <- data.frame(id = 50001:60000, value_y = rnorm(10000))
result <- bloom_join(x, y, by = "id")